+++ bd/payroll-service/app/Http/Controllers/PayrollController.php
@@ -17,34 +17,66 @@ class PayrollController extends Controller
     public function store(Request $request)
     {
         $validated = $request->validate([
-            'employee_id' => 'required|exists:employees,id',
-            'amount' => 'required|numeric',
-            'pay_date' => 'required|date',
+            'payPeriodID' => 'required|integer|exists:pay_periods,payPeriodID',
+            'employeeID' => 'required|integer|exists:employeeID',       
+                 'amount' => 'required|numeric',
+            'totalEarnings' => 'required|numeric',
+            'totalDeductions' => 'required|numeric',
+            'netpay' => 'required|numeric',
+
         ]);
 
-        return Payroll::create($validated);
+        return "hey";
     }
 
-    public function show(Payroll $payroll)
+    public function show($id)
     {
-        return $payroll->load('employee');
+        $payroll = Payroll::where('id', $id)->first();
+        if (!$payroll) {
+            return response()->json(['message' => 'not found'], 404);
+        }
+        return $payroll;
+        // return $payroll->load('employee');
     }
-
-    public function update(Request $request, Payroll $payroll)
+    public function showPayrollSpecificEmployee(Payroll $payroll)
     {
+        // $earning = Earning::where('id', $id)->first();
+        // if ($earning) {
+        //     return response()->json(['message' => 'not found'], 404);
+        // }
+        // return $earning;
+        // return $payroll->load('employee');
+    }
+    // public function update(Request $request, Payroll $payroll)
+
+
+    public function update(Request $request, $id)
+    {
+       
+        $payroll = Payroll::where('id', $id)->first();
+        if (!$payroll) {
+            return response()->json(['message' => 'not found'], 404);
+        }
         $validated = $request->validate([
-            'employee_id' => 'sometimes|exists:employees,id',
-            'amount' => 'sometimes|numeric',
-            'pay_date' => 'sometimes|date',
+            'payPeriodID' => 'required|integer|exists:payPeriodID',
+            'employeeID' => 'required|integer|exists:employeeID',            'amount' => 'required|numeric',
+            'totalEarnings' => 'required|numeric',
+            'totalDeductions' => 'required|numeric',
+            'netpay' => 'required|numeric',
         ]);
 
         $payroll->update($validated);
 
         return $payroll;
     }
+    // public function destroy(Payroll $payroll)
 
-    public function destroy(Payroll $payroll)
+    public function destroy(Payroll $id)
     {
+        $payroll = Payroll::where('id', $id)->first();
+        if (!$payroll) {
+            return response()->json(['message' => 'not found'], 404);
+        }
         $payroll->delete();
 
         return response()->noContent();
